refactor(website): create route groups
I needed to seperate layouts for `/admin` route (and it's potential children in the future) & the rest of the website.
