+++ src/app/(admin)/admin/page.tsx
@@ -1,0 +1,5 @@
+const Admin = () => (
+    <></>
+);
+
+export default Admin;
