+++ src/app/(website)/loading.tsx
@@ -1,0 +1,5 @@
+import { Loading } from "@khenzii-dev/ui/atoms";
+
+const Skeleton = () => <Loading size={200} />;
+
+export default Skeleton;
