--- src/app/admin/page.tsx
@@ -1,11 +1,0 @@
-import type { Metadata } from "next";
-
-export const metadata: Metadata = {
-    title: "Admin",
-};
-
-const Admin = () => (
-    <></>
-);
-
-export default Admin;
