+++ src/styles/layout_admin.module.scss
@@ -1,0 +1,14 @@
+@import "./constants.scss";
+
+.layout {
+    background-color: $contentBackground;
+    overflow-x: hidden;
+    margin: 0;
+}
+
+.content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
